+++ scripts/LoopSubmit.sh
@@ -21,9 +21,9 @@ for i in $(seq 1 $N); do
     export JOBID=$i  # Export jobid as an environment variable
     sbatch "$SUBDIR/$script"
     if [ "$i" -eq 1 ]; then
-        sleep 10
+        sleep 60
     else
-        sleep 1
+        sleep 10
     fi
 done